Rename MessageBus to ServiceBus; apply file-scoped namespace
