+++ EchoServiceApi/Verifiers/CosmosVerifier.cs
@@ -2,36 +2,35 @@
 using NetLah.Extensions.Configuration;
 
 #pragma warning disable S4457 // Parameter validation in "async"/"await" methods should be wrapped
-namespace EchoServiceApi.Verifiers
+namespace EchoServiceApi.Verifiers;
+
+public class CosmosVerifier : BaseCosmosVerifier
 {
-    public class CosmosVerifier : BaseCosmosVerifier
+    public CosmosVerifier(IServiceProvider serviceProvider) : base(serviceProvider) { }
+
+    public async Task<VerifyResult> VerifyAsync(string name, string key)
     {
-        public CosmosVerifier(IServiceProvider serviceProvider) : base(serviceProvider) { }
+        if (string.IsNullOrEmpty(key))
+            throw new ArgumentNullException(nameof(key));
 
-        public async Task<VerifyResult> VerifyAsync(string name, string key)
-        {
-            if (string.IsNullOrEmpty(key))
-                throw new ArgumentNullException(nameof(key));
+        var connectionObj = GetConnection(name);
+        var cosmosInfo = connectionObj.Get<CosmosContainerInfo>();
 
-            var connectionObj = GetConnection(name);
-            var cosmosInfo = connectionObj.Get<CosmosContainerInfo>();
+        var containerName = cosmosInfo.ContainerName;
+        var databaseName = cosmosInfo.DatabaseName;
 
-            var containerName = cosmosInfo.ContainerName;
-            var databaseName = cosmosInfo.DatabaseName;
+        using var cosmosclient = await CreateClientAsync(connectionObj, cosmosInfo);
 
-            using var cosmosclient = await CreateClientAsync(connectionObj, cosmosInfo);
+        using var scope = LoggerBeginScopeDiagnostic();
 
-            using var scope = LoggerBeginScopeDiagnostic();
+        Logger.LogInformation("CosmosVerifier db:{databaseName} container:{containerName} name={query_name} key={query_key}",
+            databaseName, containerName, name, key);
 
-            Logger.LogInformation("CosmosVerifier db:{databaseName} container:{containerName} name={query_name} key={query_key}",
-                databaseName, containerName, name, key);
+        var container = cosmosclient.GetContainer(databaseName, containerName);
+        _ = await container.ReadContainerAsync().ConfigureAwait(false);
 
-            var container = cosmosclient.GetContainer(databaseName, containerName);
-            _ = await container.ReadContainerAsync().ConfigureAwait(false);
-
-            var itemResponse = await container.ReadItemAsync<Dictionary<string, object>>(key, new PartitionKey(key));
-            return VerifyResult.SuccessObject("Cosmos", connectionObj, itemResponse.Resource);
-        }
+        var itemResponse = await container.ReadItemAsync<Dictionary<string, object>>(key, new PartitionKey(key));
+        return VerifyResult.SuccessObject("Cosmos", connectionObj, itemResponse.Resource);
     }
 }
 #pragma warning restore S4457 // Parameter validation in "async"/"await" methods should be wrapped
