+++ EchoServiceApi/Verifiers/DirVerifier.cs
@@ -1,70 +1,69 @@
-﻿namespace EchoServiceApi.Verifiers
+﻿namespace EchoServiceApi.Verifiers;
+
+public class DirVerifier : BaseVerifier
 {
-    public class DirVerifier : BaseVerifier
+    public DirVerifier(IServiceProvider serviceProvider) : base(serviceProvider) { }
+
+    public Task<VerifyResult> VerifyAsync(string path)
     {
-        public DirVerifier(IServiceProvider serviceProvider) : base(serviceProvider) { }
+        if (string.IsNullOrEmpty(path))
+            throw new ArgumentException("Path is required");
 
-        public Task<VerifyResult> VerifyAsync(string path)
+        var isDir = false;
+        var isFile = File.Exists(path);
+        if (!isFile)
         {
-            if (string.IsNullOrEmpty(path))
-                throw new ArgumentException("Path is required");
+            isDir = Directory.Exists(path);
+        }
 
-            var isDir = false;
-            var isFile = File.Exists(path);
-            if (!isFile)
+        if (isFile)
+        {
+            Logger.LogInformation("DirVerifier: File: path={query_path}", path);
+
+            var fileInfo = new FileInfo(path);
+            var detail = $"length={fileInfo.Length}; created={fileInfo.CreationTime}; updated={fileInfo.LastWriteTime}";
+            return Task.FromResult<VerifyResult>(new VerifySuccessMessage
             {
-                isDir = Directory.Exists(path);
+                Message = $"File '{path}' is exist",
+                Detail = detail,
+            });
+        }
+        else if (isDir)
+        {
+            Logger.LogInformation("DirVerifier: Directory: path={path}", path);
+
+            var path1 = Path.GetFullPath(path);
+            var dirInfo = new DirectoryInfo(path1);
+            var pos = path1.Length + 1;
+            var files = dirInfo.GetFiles();
+
+            var someFiles = files
+                .OrderBy(s => s.FullName)
+                .Select(s => s.FullName[pos..])
+                .Take(10)
+                .ToArray();
+
+            if (someFiles.Length < files.Length)
+            {
+                someFiles = someFiles.Append("...").ToArray();
             }
 
-            if (isFile)
+            var detail = $"length={files.Length}";
+            return Task.FromResult<VerifyResult>(new VerifySuccess<string[]>
             {
-                Logger.LogInformation("DirVerifier: File: path={query_path}", path);
-
-                var fileInfo = new FileInfo(path);
-                var detail = $"length={fileInfo.Length}; created={fileInfo.CreationTime}; updated={fileInfo.LastWriteTime}";
-                return Task.FromResult<VerifyResult>(new VerifySuccessMessage
-                {
-                    Message = $"File '{path}' is exist",
-                    Detail = detail,
-                });
-            }
-            else if (isDir)
+                Message = $"Directory '{path}' is exist",
+                Detail = detail,
+                Value = someFiles
+            });
+        }
+        else
+        {
+            return Task.FromResult<VerifyResult>(new VerifySuccessMessage
             {
-                Logger.LogInformation("DirVerifier: Directory: path={path}", path);
-
-                var path1 = Path.GetFullPath(path);
-                var dirInfo = new DirectoryInfo(path1);
-                var pos = path1.Length + 1;
-                var files = dirInfo.GetFiles();
-
-                var someFiles = files
-                    .OrderBy(s => s.FullName)
-                    .Select(s => s.FullName[pos..])
-                    .Take(10)
-                    .ToArray();
-
-                if (someFiles.Length < files.Length)
-                {
-                    someFiles = someFiles.Append("...").ToArray();
-                }
-
-                var detail = $"length={files.Length}";
-                return Task.FromResult<VerifyResult>(new VerifySuccess<string[]>
-                {
-                    Message = $"Directory '{path}' is exist",
-                    Detail = detail,
-                    Value = someFiles
-                });
-            }
-            else
-            {
-                return Task.FromResult<VerifyResult>(new VerifySuccessMessage
-                {
-                    Success = false,
-                    Message = $"Path '{path}' is not exist",
-                    Detail = null,
-                });
-            }
+                Success = false,
+                Message = $"Path '{path}' is not exist",
+                Detail = null,
+            });
         }
     }
 }
